add new modelchecker file and put shared methods there
